Add controller unit tests

+++ Deosrc.MoviesTechnicalTest.Api/Models/Responses/MovieResponse.cs
@@ -6,9 +6,9 @@ namespace Deosrc.MoviesTechnicalTest.Api.Models.Responses
         public required string Title { get; set; }
         public required string ReleaseDate { get; set; }
         public required string Overview { get; set; }
-        public required decimal Popularity { get; set; }
+        public required double Popularity { get; set; }
         public required int VoteCount { get; set; }
-        public required decimal VoteAverage { get; set; }
+        public required double VoteAverage { get; set; }
         public required string OriginalLanguage { get; set; }
         public required string[] Genres { get; set; }
         public required string PosterUrl { get; set; }
